renamed hook file + moved interface to type.ts
