+++ apps/medon-fe/src/components/AvailabilityCalendar/types.ts
@@ -1,0 +1,4 @@
+export interface SelectHours {
+  start: number;
+  end: number;
+}
